refactor(data-feeds-config-generator): Move `feed-config` types to `libs/config-types`
