--- apps/data-feeds-config-generator/src/index.ts
@@ -1,1 +1,0 @@
-export * from './feeds-config/types';
