+++ apps/docs.blocksense.network/package.json
@@ -2,7 +2,7 @@
   "name": "@blocksense/docs.blocksense.network",
   "version": "0.0.0",
   "scripts": {
-    "build:deps": "yarn workspace @blocksense/docs-theme build && yarn workspace @blocksense/contracts sol-reflect && yarn workspace @blocksense/base-utils build && yarn workspace @blocksense/data-feeds-config-generator build",
+    "build:deps": "yarn workspace @blocksense/docs-theme build && yarn workspace @blocksense/contracts sol-reflect && yarn workspace @blocksense/base-utils build && yarn workspace @blocksense/config-types build",
     "build": "yarn build:deps && yarn build:tsc && yarn generate-contracts-mdx && yarn generate-data-feed-mdx && next build",
     "build:tsc": "yarn tsc -p ./tsconfig.lib.json",
     "dev": "next",
@@ -28,8 +28,8 @@
   },
   "dependencies": {
     "@blocksense/base-utils": "workspace:*",
+    "@blocksense/config-types": "workspace:*",
     "@blocksense/contracts": "workspace:*",
-    "@blocksense/data-feeds-config-generator": "workspace:*",
     "@blocksense/docs-theme": "workspace:*",
     "@blocksense/monorepo": "workspace:*",
     "@blocksense/sol-reflector": "workspace:*",
