+++ libs/config-types/src/data-feeds-config/index.ts
