+++ libs/config-types/src/index.ts
@@ -1,3 +1,1 @@
-export function hello() {
-  return 'Hello';
-}
+export * from './data-feeds-config';
